+++ tests/testthat/test-area_of_applicability.R
@@ -3,6 +3,10 @@ skip_if_not(rlang::is_installed("vip"))
 train <- vip::gen_friedman(1000, seed = 101)
 test <- train[701:1000, ]
 train <- train[1:700, ]
+comb_rset <- rsample::make_splits(train, test)
+comb_rset <- rsample::manual_rset(list(comb_rset), "Fold1")
+comb_rset_no_y <- rsample::make_splits(train[2:11], test[2:11])
+comb_rset_no_y <- rsample::manual_rset(list(comb_rset_no_y), "Fold1")
 
 pp <- ppr(y ~ ., data = train, nterms = 11)
 importance <- vip::vi_permute(
@@ -29,12 +33,13 @@ test_that("`ww_area_of_applicability` finds 0 distance between identical data", 
 
   expect_equal(
     ww_area_of_applicability(y ~ ., train, train, importance)$aoa_threshold,
-    0
+    0,
+    tolerance = 1e-7
   )
 
 })
 
-test_that("`ww_area_of_applicability` works with or without a validation set", {
+test_that("`ww_area_of_applicability` works with or without a testing set", {
 
   expect_error(
     ww_area_of_applicability(y ~ ., train, test, importance),
@@ -53,39 +58,48 @@ test_that("`ww_area_of_applicability` methods are equivalent", {
   methods <- list(
     ww_area_of_applicability(y ~ ., train, test, importance),
     ww_area_of_applicability(train[2:11], test[2:11], importance),
-    ww_area_of_applicability(as.matrix(train[2:11]), as.matrix(test[2:11]), importance)
+    ww_area_of_applicability(as.matrix(train[2:11]), as.matrix(test[2:11]), importance),
+    ww_area_of_applicability(comb_rset_no_y, importance = importance)
   )
 
   expect_identical(
-    methods[[1]]$aoa_threshold,
-    methods[[2]]$aoa_threshold
+    head(methods[[1]], -1),
+    head(methods[[2]], -1)
   )
 
   expect_identical(
-    methods[[2]]$aoa_threshold,
-    methods[[3]]$aoa_threshold
+    head(methods[[2]], -1),
+    head(methods[[3]], -1)
+  )
+
+  # Comparing rset method to the others --
+  # because here we calculate our training data on the entire thing
+  # the training, means, sds slots are all different
+  expect_identical(
+    methods[[3]]$aoa_threshold,
+    methods[[4]]$aoa_threshold
+  )
+
+  skip_if_not_installed("recipes")
+  expect_identical(
+    head(methods[[4]], -1),
+    head(ww_area_of_applicability(comb_rset, recipes::recipe(y ~ ., train), importance = importance), -1)
   )
 
 })
 
 test_that("`ww_area_of_applicability` can handle different column orders", {
 
-  expect_identical(
+  expect_equal(
     ww_area_of_applicability(train[2:11], test[2:11], importance)$aoa_threshold,
     ww_area_of_applicability(train[2:11], test[11:2], importance)$aoa_threshold
   )
 
-  expect_identical(
+  expect_equal(
     ww_area_of_applicability(train[2:11], test[2:11], importance)$aoa_threshold,
     ww_area_of_applicability(train[11:2], test[2:11], importance)$aoa_threshold
   )
 
-})
-
-test_that("`predict_ww_area_of_applicability_numeric` fails when model has no pcs argument", {
-  expect_snapshot(error = TRUE,
-                  predict_ww_area_of_applicability_numeric(mtcars, mtcars)
-  )
 })
 
 skip_if_not(rlang::is_installed("vip"))
@@ -118,12 +132,15 @@ test_that("normal use", {
 test_that("`new_ww_area_of_applicability` arguments are assigned correctly", {
   x <- ww_area_of_applicability(y ~ ., train, test, importance)
 
+  expect_identical(
+    as.data.frame(x$training),
+    train[2:11]
+  )
+
   skip_on_os("mac")
-  expect_equal(names(x), c("training", "importance", "sds", "means", "d_bar", "aoa_threshold", "blueprint"))
+  expect_equal(names(x), c("training", "importance", "d_bar", "aoa_threshold", "blueprint"))
   expect_snapshot(x$training)
   expect_snapshot(x$importance)
-  expect_snapshot(x$sds)
-  expect_snapshot(x$means)
   expect_snapshot(x$d_bar)
   expect_snapshot(x$aoa_threshold)
   expect_snapshot(x$blueprint)