+++ NEWS.md
@@ -1,5 +1,7 @@
 # waywiser (development version)
 
+* Added functions for calculating the area of applicability of a model. 
+
 # waywiser 0.1.0
 
 * Added functions for automatically constructing `nb` and `listw` objects